Use custom Behat feature for WordPress login (#79)
It turns out the Behat issue was not with GitLab but with the WordHat. I posted my login woes in their Slack instance and got this reply

> The log-in step has been always troublesome. If you look at the browser at the failing test, is it doing something like entering the username but not the password - something like that?
> If it is it is probably the project’s oldest bug and several of us have been unable to reliably fix it.

So I spent a day writing custom features to make login more reliable and have some Pantheon specific items as well.

The upstream changes were merged into WordHat and released in `3.1.1`.

In addition, I think showing off custom contexts will be helpful to people learning Behat.